refactor: Refactor PrepareFaucet integration test
